Moved implementation into sion-vector.cpp.
